Update brave.json
Realized the scoop version of Brave that this script linked to on github is a version not compatible with Brave  wallet creation and use of Basic Attention Token. I changed this script so both the 64bit and 32bit versions are pulled from the links on Brave's homepage so scoop users can install and update the current stable release. However, I'm not sure that those hash numbers will work with the correct versions. How would I go about updating the hashes properly?